+++ test/test.js
@@ -1,5 +1,9 @@
-import Typeson from '../typeson.js';
-import * as B64 from 'base64-arraybuffer-es6';
+import '../node_modules/core-js-bundle/minified.js';
+import '../node_modules/regenerator-runtime/runtime.js';
+
+import Typeson from '../dist/typeson-esm.js';
+import * as B64 from
+    '../node_modules/base64-arraybuffer-es6/dist/base64-arraybuffer-es.js';
 
 const typeson = new Typeson().register({
     Date: [