+++ utils/TypesonPromise.js
@@ -9,7 +9,7 @@
 class TypesonPromise{constructor(f){this.p=new Promise(f)}} // eslint-disable-line block-spacing, space-before-function-paren, space-before-blocks, space-infix-ops, semi
 // class TypesonPromise extends Promise {get[Symbol.toStringTag](){return 'TypesonPromise'};} // eslint-disable-line keyword-spacing, space-before-function-paren, space-before-blocks, block-spacing, semi
 
-// Note: @babel/polyfill provides a `Symbol` polyfill
+// Note: core-js-bundle provides a `Symbol` polyfill
 if (typeof Symbol !== 'undefined') {
     // Ensure `isUserObject` will return `false` for `TypesonPromise`
     TypesonPromise.prototype[Symbol.toStringTag] = 'TypesonPromise';